currency service update and some additional changes

+++ account-service/src/main/java/com/sau/swe/dto/AccountResponse.java
@@ -4,17 +4,18 @@ import com.sau.swe.entity.Account;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @NoArgsConstructor
 @Data
 public class AccountResponse {
-    private Long balance;
+    private BigDecimal balance;
     private String accountType;
     private String transferCode;
     private List<TransactionDTO> lastTransactions;
 
-    public AccountResponse(Long balance, Integer accountType, String transferCode) {
+    public AccountResponse(BigDecimal balance, Integer accountType, String transferCode) {
         this.balance = balance;
         this.accountType = Account.AccountType.values()[accountType].name();
         this.transferCode = transferCode;
